+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+
+## [0.1.4] - 2017-06-11
 ### Changed
 - Renamed `create:*` commands to `make:*` ([#5](https://github.com/svenluijten/forge-cli/issues/5))
 
@@ -72,7 +74,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `update:site` command
 - `delete:site` command
 
-[Unreleased]: https://github.com/svenluijten/forge-cli/compare/0.1.3...HEAD
+[Unreleased]: https://github.com/svenluijten/forge-cli/compare/0.1.4...HEAD
+[0.1.4]: https://github.com/svenluijten/forge-cli/compare/0.1.3...0.1.4
 [0.1.3]: https://github.com/svenluijten/forge-cli/compare/0.1.2...0.1.3 
 [0.1.2]: https://github.com/svenluijten/forge-cli/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/svenluijten/forge-cli/compare/0.1.0...0.1.1